Generalize functions to apply to stacks of images
By adding a axes parameter to all flavors of the per operator.
Note: This still only applies the transform to 2D slices of the nD array
and is no true generalization to N-dimensional datasets (the original
paper does not treat the nD case, although thge extension seems
straightforward to implement.